requirements.txt
Django>=5.2.8
python-decouple>=3.8
django-grappelli>=3.1.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
django-filter>=23.0

